refactor(general) : MP-318 Change Splash Screen Method
