+++ app/src/main/AndroidManifest.xml
@@ -17,26 +17,23 @@
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/logo_nuberjam_title_rounded"
         android:label="@string/app_name"
         android:requestLegacyExternalStorage="true"
         android:supportsRtl="true"
-        android:theme="@style/Theme.NuberJam"
+        android:theme="@style/Theme.AppSplashScreen"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SplashScreenActivity"
+            android:name=".ui.authentication.AuthActivity"
             android:exported="true"
-            android:theme="@style/SplashTheme">
+            android:launchMode="singleTop">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ui.authentication.AuthActivity"
-            android:exported="true"
-            android:launchMode="singleTop"/>
         <activity
             android:name=".ui.main.MainActivity"
             android:exported="true"
