+++ app/src/main/java/com/example/nuberjam/ui/SplashViewModel.kt
@@ -4,5 +4,6 @@ import androidx.lifecycle.ViewModel
 import com.example.nuberjam.data.Repository
 
 class SplashViewModel (private val repository: Repository) : ViewModel() {
+    var isLoading = true
     fun getLoginState() = repository.getLoginState()
 }
