+++ app/src/main/res/values/themes.xml
@@ -57,7 +57,4 @@
         <item name="android:fontFamily">@font/century_gothic_bold</item>
     </style>
 
-    <style name="SplashTheme" parent="Theme.NuberJam">
-        <item name="android:windowBackground">@drawable/splash_background</item>
-    </style>
 </resources>
